requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Django==4.1.3
django-dotenv==1.4.2
djangorestframework==3.14.0
jwt==1.3.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
